Add Authorize checks to ApplicationHelper,
for use by controllers or other parts of app.
Updated authenticate function for User to deal with bcrypt password
Added custom setter for Password in User
At this point, we can se the User class to add a user to the database;
the mass assignment does not work however.
To add a user - do the following in the rails console
newuser = User.new
newuser.UserID = 'userid'
newuser.Password = 'cleartextpassword' # User model will bcrypt it
newuser.UserName = 'New User Name'
newuser.AuthorizationLevel = 2 # or 1, or 3, or 4
newuser.FaildAttempts = 0
newuser.save

To get the user out of the system:
newuser = User.authenticate(userid, cleartextpassword) # we use bcrypt
to test for match