nav structure  changes
We now have views for student, teacher and public under math_content.

The intent is to create full views for each page type, and have the
dynamic content run at teh view level.  While this is not Rails best
practice, it's wasier to think about for a first Rails project.  Once
we understand it better, we can refactor and move things up to the
controllers.